+++ code/Compress/Lzo/InflateStreamLzo.cpp
@@ -27,7 +27,7 @@ public:
 		m_stream = 0;
 	}
 
-	int read(void* block, int nbytes)
+	int64_t read(void* block, int64_t nbytes)
 	{
 		uint8_t* top = static_cast< uint8_t* >(block);
 		uint8_t* ptr = static_cast< uint8_t* >(block);
@@ -35,7 +35,7 @@ public:
 		// Copy from buffer.
 		if (m_decompressedBufferSize > 0)
 		{
-			int32_t ncopy = std::min< int32_t >(m_decompressedBufferSize, nbytes);
+			int64_t ncopy = std::min< int64_t >(m_decompressedBufferSize, nbytes);
 
 			std::memcpy(ptr, &m_decompressedBuffer[0], ncopy);
 			ptr += ncopy;
@@ -49,15 +49,15 @@ public:
 
 		if (nbytes <= 0)
 		{
-			m_position += int32_t(ptr - top);
-			return int32_t(ptr - top);
+			m_position += int64_t(ptr - top);
+			return int64_t(ptr - top);
 		}
 
 		T_ASSERT (m_decompressedBufferSize == 0);
 
 		// Decompress directly into destination if requested block size is larger than buffer;
 		// and output is large enough to contain a bit of overhead.
-		while (nbytes >= int(m_decompressedBuffer.size()))
+		while (nbytes >= int64_t(m_decompressedBuffer.size()))
 		{
 			uint32_t compressedBlockSize = 0;
 			Reader(m_stream) >> compressedBlockSize;
@@ -109,7 +109,7 @@ public:
 			if (r != LZO_E_OK || !m_decompressedBufferSize)
 				break;
 
-			int32_t ncopy = std::min< int32_t >(m_decompressedBufferSize, nbytes);
+			int64_t ncopy = std::min< int64_t >(m_decompressedBufferSize, nbytes);
 
 			std::memcpy(ptr, &m_decompressedBuffer[0], ncopy);
 			ptr += ncopy;
@@ -121,11 +121,11 @@ public:
 			nbytes -= ncopy;
 		}
 
-		m_position += int32_t(ptr - top);
-		return int32_t(ptr - top);
+		m_position += int64_t(ptr - top);
+		return int64_t(ptr - top);
 	}
 
-	int setLogicalPosition(int position)
+	int64_t setLogicalPosition(int64_t position)
 	{
 		// Seeking backwards, restart from beginning.
 		if (position < m_position)
@@ -138,7 +138,7 @@ public:
 		uint8_t dummy[1024];
 		while (m_position < position)
 		{
-			int nread = read(dummy, std::min< int >(sizeof_array(dummy), position - m_position));
+			int64_t nread = read(dummy, std::min< int64_t >(sizeof_array(dummy), position - m_position));
 			if (nread <=  0)
 				return -1;
 		}
@@ -146,7 +146,7 @@ public:
 		return m_position;
 	}
 
-	int getLogicalPosition() const
+	int64_t getLogicalPosition() const
 	{
 		return m_position;
 	}
@@ -155,9 +155,9 @@ private:
 	Ref< IStream > m_stream;
 	AlignedVector< uint8_t > m_compressedBlock;
 	AlignedVector< uint8_t > m_decompressedBuffer;
-	int32_t m_decompressedBufferSize;
-	int32_t m_startPosition;
-	int32_t m_position;
+	int64_t m_decompressedBufferSize;
+	int64_t m_startPosition;
+	int64_t m_position;
 };
 
 T_IMPLEMENT_RTTI_CLASS(L"traktor.compress.InflateStreamLzo", InflateStreamLzo, IStream)
@@ -196,18 +196,18 @@ bool InflateStreamLzo::canSeek() const
 	return true;
 }
 
-int InflateStreamLzo::tell() const
+int64_t InflateStreamLzo::tell() const
 {
 	return m_impl->getLogicalPosition();
 }
 
-int InflateStreamLzo::available() const
+int64_t InflateStreamLzo::available() const
 {
 	T_FATAL_ERROR;
 	return 0;
 }
 
-int InflateStreamLzo::seek(SeekOriginType origin, int offset)
+int64_t InflateStreamLzo::seek(SeekOriginType origin, int64_t offset)
 {
 	T_ASSERT_M (origin != SeekEnd, L"SeekEnd is not allowed");
 	if (origin == SeekCurrent)
@@ -215,12 +215,12 @@ int InflateStreamLzo::seek(SeekOriginType origin, int offset)
 	return m_impl->setLogicalPosition(offset);
 }
 
-int InflateStreamLzo::read(void* block, int nbytes)
+int64_t InflateStreamLzo::read(void* block, int64_t nbytes)
 {
 	return m_impl->read(block, nbytes);
 }
 
-int InflateStreamLzo::write(const void* block, int nbytes)
+int64_t InflateStreamLzo::write(const void* block, int64_t nbytes)
 {
 	T_FATAL_ERROR;
 	return 0;
